+++ .flutter-plugins
@@ -1,2 +1,2 @@
 # This is a generated file; do not edit or check into version control.
-flutter_email_sender=C:\\Users\\cathy\\AppData\\Local\\Pub\\Cache\\hosted\\pub.dartlang.org\\flutter_email_sender-5.1.0\\
+flutter_email_sender=C:\\src\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\flutter_email_sender-5.1.0\\
