+++ .idea/libraries/Flutter_Plugins.xml
@@ -1,7 +1,7 @@
 <component name="libraryTable">
   <library name="Flutter Plugins" type="FlutterPluginsLibraryType">
     <CLASSES>
-      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dartlang.org/flutter_email_sender-5.1.0" />
+      <root url="file://$PROJECT_DIR$/../../../../src/flutter/.pub-cache/hosted/pub.dartlang.org/flutter_email_sender-5.1.0" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
